Cross-build for sbt 0.13.x and 1.x
