+++ build.sbt
@@ -1,6 +1,7 @@
 import ReleaseTransformations._
 
 enablePlugins(SbtPlugin)
+crossSbtVersions := Seq("0.13.17", "1.2.3")
 
 scalaVersion := "2.12.6"
 organization := "com.github.cb372"
