+++ src/main/scala-sbt-1.0/explicitdeps/package.scala
@@ -1,0 +1,8 @@
+package object explicitdeps {
+
+  type Logger = sbt.util.Logger
+  type ModuleID = sbt.librarymanagement.ModuleID
+  type Binary = sbt.librarymanagement.Binary
+  type Analysis = sbt.internal.inc.Analysis
+
+}
