+++ src/main/scala/explicitdeps/UndeclaredCompileDependenciesException.scala
@@ -1,0 +1,5 @@
+package explicitdeps
+
+object UndeclaredCompileDependenciesException extends sbt.FeedbackProvidedException {
+  override def toString: String = "Failing the build because undeclared dependencies were found"
+}
