+++ src/main/scala/explicitdeps/UnusedCompileDependenciesException.scala
@@ -1,0 +1,5 @@
+package explicitdeps
+
+object UnusedCompileDependenciesException extends sbt.FeedbackProvidedException {
+  override def toString: String = "Failing the build because unused dependencies were found"
+}
